ISS-87. Provided support for optionally ignoring -ktx dependencies.
