+++ src/main/kotlin/com/autonomousapps/internal/advice/Advisor.kt
@@ -1,12 +1,14 @@
 package com.autonomousapps.internal.advice
 
-import com.autonomousapps.Behavior
-import com.autonomousapps.Ignore
-import com.autonomousapps.Warn
-import com.autonomousapps.internal.*
+import com.autonomousapps.internal.AnalyzedJar
+import com.autonomousapps.internal.Component
+import com.autonomousapps.internal.Dependency
+import com.autonomousapps.internal.TransitiveComponent
+import com.autonomousapps.internal.UnusedDirectComponent
 import com.autonomousapps.internal.utils.filterToOrderedSet
+import com.autonomousapps.internal.utils.filterToSet
 import com.autonomousapps.internal.utils.mapToOrderedSet
-import org.gradle.api.GradleException
+import com.autonomousapps.internal.utils.mapToSet
 
 /**
  * Classes of advice:
@@ -15,13 +17,18 @@ import org.gradle.api.GradleException
  * 3. `implementation` dependencies incorrectly declared `api`
  * 4. `api` dependencies incorrectly declared `implementation`.
  * 5. `compileOnly` dependencies incorrectly declared as `api` or `implementation`.
+ *
+ * [ignoreKtx] is applied to the remove-advice to filter (or not) ktx-related advice; supplied by
+ * user. If true, we don't suggest removing an "unused" `-ktx` dependency iff any of its
+ * dependencies are used (directly or transitively).
  */
 internal class Advisor(
   private val allComponents: List<Component>,
   private val unusedDirectComponents: List<UnusedDirectComponent>,
   private val usedTransitiveComponents: List<TransitiveComponent>,
   private val abiDeps: List<Dependency>,
-  private val allDeclaredDeps: List<Dependency>
+  private val allDeclaredDeps: List<Dependency>,
+  private val ignoreKtx: Boolean = false
 ) {
 
   /**
@@ -31,7 +38,8 @@ internal class Advisor(
     universalFilter = CompositeFilter(listOf(DataBindingFilter(), ViewBindingFilter()))
   )): ComputedAdvice {
     val compileOnlyCandidates = computeCompileOnlyCandidates()
-    val adviceUnused = computeUnusedDependencies(compileOnlyCandidates)
+    val unusedDependencies = computeUnusedDependencies(compileOnlyCandidates)
+    val ktxDependencies = computeKtxDependencies(unusedDependencies)
     val undeclaredApiDependencies = computeUndeclaredApiDependencies(compileOnlyCandidates)
     val undeclaredImplDependencies = computeUndeclaredImplDependencies(undeclaredApiDependencies, compileOnlyCandidates)
     val changeToApi = computeApiDepsWronglyDeclared(compileOnlyCandidates)
@@ -39,7 +47,8 @@ internal class Advisor(
 
     return ComputedAdvice(
       compileOnlyCandidates = compileOnlyCandidates,
-      unusedDependencies = adviceUnused,
+      unusedDependencies = unusedDependencies,
+      ktxDependencies = ktxDependencies,
       undeclaredApiDependencies = undeclaredApiDependencies,
       undeclaredImplDependencies = undeclaredImplDependencies,
       changeToApi = changeToApi,
@@ -77,6 +86,40 @@ internal class Advisor(
   }
 
   /**
+   * "KTX Dependencies" are those which
+   * 1. Have an [identifier][Dependency.identifier] that ends with "-ktx"
+   * 2. are nominally unused by which contribute transitive dependencies which _are_ used. We only
+   *    care about them if [FilterSpec.ignoreKtx] is `true`.
+   *
+   * @return the set of dependencies which must be filtered from the set returned by
+   * [computeUnusedDependencies], to respect the user's preference to "ignore ktx dependencies."
+   */
+  private fun computeKtxDependencies(unusedDependencies: Set<Dependency>): Set<Dependency> {
+    if (!ignoreKtx) {
+      return emptySet()
+    }
+
+    val usedDependencies: Set<Dependency> = allComponents
+      .filterToSet { !it.isTransitive }
+      .mapToSet { it.dependency }
+      .filterToSet { directDependency ->
+        unusedDependencies.none {
+          it == directDependency
+        }
+      } + usedTransitiveComponents.mapToSet { it.dependency }
+
+    // Filter out those that are used, whether they've been declared directly or are brought in
+    // transitively.
+    return unusedDependencies
+      .filterToOrderedSet { it.identifier.endsWith("-ktx") }
+      .filterToOrderedSet { unusedDependency ->
+        usedDependencies.none {
+          it == unusedDependency
+        }
+      }
+  }
+
+  /**
    * A [Dependency] is an undeclared `api` dependency (and should be declared as such) iff:
    * 1. It is part of the project's ABI AND
    * 2. It was not declared (it's [configurationName][Dependency.configurationName] is `null`) AND
@@ -86,19 +129,6 @@ internal class Advisor(
     return abiDeps
       .filterToOrderedSet { it.configurationName == null }
       .stripCompileOnly(compileOnlyCandidates)
-//      .filterToOrderedSet { dep ->
-//        compileOnlyCandidates.none { compileOnly ->
-//          dep == compileOnly.dependency
-//        }
-//      }
-  }
-
-  private fun Iterable<Dependency>.stripCompileOnly(compileOnlyCandidates: Set<Component>): Set<Dependency> {
-    return filterToOrderedSet { dep ->
-      compileOnlyCandidates.none { compileOnly ->
-        dep == compileOnly.dependency
-      }
-    }
   }
 
   /**
@@ -119,12 +149,7 @@ internal class Advisor(
           api == trans
         }
       }
-      // Don't suggest adding a compileOnly candidate. They're handled elsewhere.
-      .filterToOrderedSet { dep ->
-        compileOnlyCandidates.none { compileOnlyComponent ->
-          compileOnlyComponent.dependency == dep
-        }
-      }
+      .stripCompileOnly(compileOnlyCandidates)
   }
 
   /**
@@ -139,17 +164,14 @@ internal class Advisor(
       .filterToOrderedSet { it.configurationName != null }
       // Filter out those with an "api" configuration, as they're already correct.
       .filterToOrderedSet { !it.configurationName!!.endsWith("api", ignoreCase = true) }
-      // Don't suggest adding a compileOnly candidate. They're handled elsewhere.
-      .filterToOrderedSet { dep ->
-        compileOnlyCandidates.none { compileOnlyComponent ->
-          compileOnlyComponent.dependency == dep
-        }
-      }
+      .stripCompileOnly(compileOnlyCandidates)
   }
 
   /**
    * A [Dependency] is a "wrongly declared" impl dep (and should be changed) iff:
-   * 1.
+   * 1. It is not transitive ([configuration][Dependency.configurationName] must be non-null).
+   * 2. It _should_ be on `implementation`, but is on something else AND
+   * 3. It is not a `compileOnly` candidate (see [computeCompileOnlyCandidates]).
    */
   private fun computeImplDepsWronglyDeclared(
     compileOnlyCandidates: Set<Component>
@@ -165,257 +187,14 @@ internal class Advisor(
           abi == dep
         }
       }
-      // Don't suggest adding a compileOnly candidate. They're handled elsewhere.
-      .filterToOrderedSet { dep ->
-        compileOnlyCandidates.none { compileOnlyComponent ->
-          dep == compileOnlyComponent.dependency
-        }
+      .stripCompileOnly(compileOnlyCandidates)
+  }
+
+  private fun Iterable<Dependency>.stripCompileOnly(compileOnlyCandidates: Set<Component>): Set<Dependency> {
+    return filterToOrderedSet { dep ->
+      compileOnlyCandidates.none { compileOnly ->
+        dep == compileOnly.dependency
       }
-  }
-}
-
-internal class ComputedAdvice(
-  unusedDependencies: Set<Dependency>,
-  undeclaredApiDependencies: Set<Dependency>,
-  undeclaredImplDependencies: Set<Dependency>,
-  changeToApi: Set<Dependency>,
-  changeToImpl: Set<Dependency>,
-  filterSpec: FilterSpec,
-  compileOnlyCandidates: Set<Component>
-) {
-
-  val compileOnlyDependencies = compileOnlyCandidates
-    // We want to exclude transitives here. In other words, don't advise people to declare used-transitive components.
-    .filterToOrderedSet { !it.isTransitive }
-    .mapToOrderedSet { it.dependency }
-    .filterToOrderedSet(filterSpec.compileOnlyAdviceFilter)
-
-  val filterRemove = filterSpec.filterRemove
-  val filterAdd = filterSpec.filterAdd
-  val filterChange = filterSpec.filterChange
-  val filterCompileOnly = filterSpec.filterCompileOnly
-
-  val addToApiAdvice: Set<Advice> = undeclaredApiDependencies
-    .filterToOrderedSet(filterSpec.addAdviceFilter)
-    .mapToOrderedSet { Advice.add(it, "api") }
-
-  val addToImplAdvice: Set<Advice> = undeclaredImplDependencies
-    .filterToOrderedSet(filterSpec.addAdviceFilter)
-    .mapToOrderedSet { Advice.add(it, "implementation") }
-
-  val removeAdvice: Set<Advice> = unusedDependencies
-    .filterToOrderedSet(filterSpec.removeAdviceFilter)
-    .mapToOrderedSet { Advice.remove(it) }
-
-  val changeToApiAdvice: Set<Advice> = changeToApi
-    .filterToOrderedSet(filterSpec.changeAdviceFilter)
-    .mapToOrderedSet { Advice.change(it, toConfiguration = "api") }
-
-  val changeToImplAdvice: Set<Advice> = changeToImpl
-    .filterToOrderedSet(filterSpec.changeAdviceFilter)
-    .mapToOrderedSet { Advice.change(it, toConfiguration = "implementation") }
-
-  val compileOnlyAdvice: Set<Advice> = compileOnlyDependencies
-    .filterToOrderedSet(filterSpec.compileOnlyAdviceFilter)
-    // TODO be variant-aware
-    .mapToOrderedSet { Advice.compileOnly(it, "compileOnly") }
-
-  fun getAdvices(): Set<Advice> {
-    val advices = sortedSetOf<Advice>()
-
-    /*
-     * Doing this all in a "functional" way would result in many many intermediate sets being created, needlessly.
-     */
-
-    addToApiAdvice.forEach { advices.add(it) }
-    addToImplAdvice.forEach { advices.add(it) }
-    removeAdvice.forEach { advices.add(it) }
-    changeToApiAdvice.forEach { advices.add(it) }
-    changeToImplAdvice.forEach { advices.add(it) }
-    compileOnlyAdvice.forEach { advices.add(it) }
-
-    return advices
-  }
-
-  fun advicePrinter(): AdvicePrinter = AdvicePrinter(this)
-}
-
-/**
- * Only concerned with human-readable advice meant to be printed to the console.
- */
-internal class AdvicePrinter(private val computedAdvice: ComputedAdvice) {
-
-  /**
-   * Returns "add-advice" (or null if none) for printing to console.
-   */
-  fun getAddAdvice(): String? {
-    val undeclaredApiDeps = computedAdvice.addToApiAdvice
-    val undeclaredImplDeps = computedAdvice.addToImplAdvice
-
-    if (undeclaredApiDeps.isEmpty() && undeclaredImplDeps.isEmpty()) {
-      return null
-    }
-
-    val apiAdvice = undeclaredApiDeps.joinToString(prefix = "- ", separator = "\n- ") {
-      "api(${printableIdentifier(it.dependency)})"
-    }
-    val implAdvice = undeclaredImplDeps.joinToString(prefix = "- ", separator = "\n- ") {
-      "implementation(${printableIdentifier(it.dependency)})"
-    }
-
-    return if (undeclaredApiDeps.isNotEmpty() && undeclaredImplDeps.isNotEmpty()) {
-      "$apiAdvice\n$implAdvice"
-    } else if (undeclaredApiDeps.isNotEmpty()) {
-      apiAdvice
-    } else if (undeclaredImplDeps.isNotEmpty()) {
-      implAdvice
-    } else {
-      // One or the other list must be non-empty
-      throw GradleException("Impossible")
-    }
-  }
-
-  /**
-   * Returns "remove-advice" (or null if none) for printing to console.
-   */
-  fun getRemoveAdvice(): String? {
-    val unusedDependencies = computedAdvice.removeAdvice
-
-    if (unusedDependencies.isEmpty()) {
-      return null
-    }
-
-    return unusedDependencies.joinToString(prefix = "- ", separator = "\n- ") {
-      "${it.fromConfiguration}(${printableIdentifier(it.dependency)})"
-    }
-  }
-
-  /**
-   * Returns "change-advice" (or null if none) for printing to console.
-   */
-  fun getChangeAdvice(): String? {
-    val changeToApi = computedAdvice.changeToApiAdvice
-    val changeToImpl = computedAdvice.changeToImplAdvice
-
-    if (changeToApi.isEmpty() && changeToImpl.isEmpty()) {
-      return null
-    }
-
-    val apiAdvice = changeToApi.joinToString(prefix = "- ", separator = "\n- ") {
-      "api(${printableIdentifier(it.dependency)}) // was ${it.fromConfiguration}"
-    }
-    val implAdvice = changeToImpl.joinToString(prefix = "- ", separator = "\n- ") {
-      "implementation(${printableIdentifier(it.dependency)}) // was ${it.fromConfiguration}"
-    }
-    return if (changeToApi.isNotEmpty() && changeToImpl.isNotEmpty()) {
-      "$apiAdvice\n$implAdvice"
-    } else if (changeToApi.isNotEmpty()) {
-      apiAdvice
-    } else if (changeToImpl.isNotEmpty()) {
-      implAdvice
-    } else {
-      // One or the other list must be non-empty
-      throw GradleException("Impossible")
-    }
-  }
-
-  /**
-   * Returns "compileOnly-advice" (or null if none) for printing to console.
-   */
-  fun getCompileOnlyAdvice(): String? {
-    val compileOnlyDependencies = computedAdvice.compileOnlyDependencies
-
-    if (compileOnlyDependencies.isEmpty()) {
-      return null
-    }
-
-    return compileOnlyDependencies.joinToString(prefix = "- ", separator = "\n- ") {
-      // TODO be variant-aware
-      "compileOnly(${printableIdentifier(it)}) // was ${it.configurationName}"
-    }
-  }
-
-  private fun printableIdentifier(dependency: Dependency): String =
-    if (dependency.identifier.startsWith(":")) {
-      "project(\"${dependency.identifier}\")"
-    } else {
-      "\"${dependency.identifier}:${dependency.resolvedVersion}\""
-    }
-}
-
-/**
- * A container for the various filters, to be applied to the final advice:
- * - [universalFilter] applied to all advice; built into plugin.
- * - [anyBehavior] applied to all advice; supplied by user.
- * - [unusedDependenciesBehavior] applied to remove-advice; supplied by user.
- * - [usedTransitivesBehavior] applied to add-dependencies advice; supplied by user.
- * - [incorrectConfigurationsBehavior] applied change-advice; supplied by user.
- * - [compileOnlyBehavior] applied to compileOnly-advice; supplied by user.
- */
-internal class FilterSpec(
-  private val universalFilter: DependencyFilter = CompositeFilter(listOf(DataBindingFilter(), ViewBindingFilter())),
-  private val anyBehavior: Behavior = Warn(),
-  private val unusedDependenciesBehavior: Behavior = Warn(),
-  private val usedTransitivesBehavior: Behavior = Warn(),
-  private val incorrectConfigurationsBehavior: Behavior = Warn(),
-  private val compileOnlyBehavior: Behavior = Warn()
-) {
-
-  private val shouldIgnoreAll = anyBehavior is Ignore
-  val filterRemove = shouldIgnoreAll || unusedDependenciesBehavior is Ignore
-  val filterAdd = shouldIgnoreAll || usedTransitivesBehavior is Ignore
-  val filterChange = shouldIgnoreAll || incorrectConfigurationsBehavior is Ignore
-  val filterCompileOnly = shouldIgnoreAll || compileOnlyBehavior is Ignore
-
-  val removeAdviceFilter: (Dependency) -> Boolean = { dependency ->
-    if (anyBehavior is Ignore || unusedDependenciesBehavior is Ignore) {
-      // If we're ignoring everything or just ignoring all unused dependencies, then do that
-      false
-    } else if (anyBehavior.filter.plus(unusedDependenciesBehavior.filter).contains(dependency.identifier)) {
-      // If we're ignoring some specific dependencies, then do that
-      false
-    } else {
-      // If the dependency is universally ignored, do that
-      universalFilter.predicate(dependency)
-    }
-  }
-
-  val addAdviceFilter: (Dependency) -> Boolean = { dependency ->
-    if (anyBehavior is Ignore || usedTransitivesBehavior is Ignore) {
-      // If we're ignoring everything or just ignoring all undeclared transitive dependencies, then do that
-      false
-    } else if (anyBehavior.filter.plus(usedTransitivesBehavior.filter).contains(dependency.identifier)) {
-      // If we're ignoring some specific dependencies, then do that
-      false
-    } else {
-      // If the dependency is universally ignored, do that
-      universalFilter.predicate(dependency)
-    }
-  }
-
-  val changeAdviceFilter: (Dependency) -> Boolean = { dependency ->
-    if (anyBehavior is Ignore || incorrectConfigurationsBehavior is Ignore) {
-      // If we're ignoring everything or just ignoring all incorrectly-declared dependencies, then do that
-      false
-    } else if (anyBehavior.filter.plus(incorrectConfigurationsBehavior.filter).contains(dependency.identifier)) {
-      // If we're ignoring some specific dependencies, then do that
-      false
-    } else {
-      // If the dependency is universally ignored, do that
-      universalFilter.predicate(dependency)
-    }
-  }
-
-  val compileOnlyAdviceFilter: (Dependency) -> Boolean = { dependency ->
-    if (anyBehavior is Ignore || compileOnlyBehavior is Ignore) {
-      // If we're ignoring everything or just ignoring all compileOnly dependencies, then do that
-      false
-    } else if (anyBehavior.filter.plus(compileOnlyBehavior.filter).contains(dependency.identifier)) {
-      // If we're ignoring some specific dependencies, then do that
-      false
-    } else {
-      // If the dependency is universally ignored, do that
-      universalFilter.predicate(dependency)
     }
   }
 }
